requirements.txt
flask
spacy
nltk
swarm @ git+https://github.com/openai/swarm.git
python-dotenv
transformers
torch
gunicorn